split qwen36 moe telemetry
